Centralize turn types in a dedicated module
- Keep turn outcomes, reports, tool activity, and errors together in `turn.rs`.
- Preserve the existing public exports while keeping harness orchestration focused on execution.
